+++ src/tests/deploy/script.spec.ts
@@ -41,7 +41,7 @@ describe('script steps', () => {
 
   it("should advance if the script exits with zero code", async  () => {
     (execSync as jest.Mock<typeof import("child_process").execSync>).mockReturnValueOnce(``)
-    await expect(executeScriptPhase(deploy, metatxn, {})).resolves.toBeUndefined();
+    await expect(executeScriptPhase(deploy, metatxn, {nonInteractive: false, defaultArgs: {}})).resolves.toBeUndefined();
     expect(deploy._.phase).toEqual(`complete`);
     expect(metatxn.commit).toHaveBeenCalled();
   })
@@ -51,19 +51,19 @@ describe('script steps', () => {
       (execSync as jest.Mock).mockImplementationOnce(() => {
         throw staticErr
       })
-      await expect(executeScriptPhase(deploy, metatxn, {})).rejects.toThrow(`The deploy halted: Script src/tests/example/1-eoa.s.sol failed.`);
+      await expect(executeScriptPhase(deploy, metatxn, {nonInteractive: false, defaultArgs: {}})).rejects.toThrow(`The deploy halted: Script src/tests/example/1-eoa.s.sol failed.`);
       expect(deploy._.phase).toEqual(`script_run`);
       expect(metatxn.commit).toHaveBeenCalled();
     })
     it("the script doesn't exist", async () => {
       deploy._.segments[deploy._.segmentId].filename = `unknown_file`;
-      await expect(executeScriptPhase(deploy, metatxn, {})).rejects.toThrow(`The deploy halted: Script src/tests/example/unknown_file does not exist.`);
+      await expect(executeScriptPhase(deploy, metatxn, {nonInteractive: false, defaultArgs: {}})).rejects.toThrow(`The deploy halted: Script src/tests/example/unknown_file does not exist.`);
       expect(deploy._.phase).toEqual(`script_run`);
       expect(metatxn.commit).not.toHaveBeenCalled();
     })
     it("the env doesn't exist", async () => {
       deploy._.env = `unknownEnv`;
-      await expect(executeScriptPhase(deploy, metatxn, {})).rejects.toThrow(`No such environment: unknownEnv`);
+      await expect(executeScriptPhase(deploy, metatxn, {nonInteractive: false, defaultArgs: {}})).rejects.toThrow(`No such environment: unknownEnv`);
       expect(deploy._.phase).toEqual(`script_run`);
       expect(metatxn.commit).not.toHaveBeenCalled();
     })